bk - reverted to working LinkController from master
Previously used a copied and pasted, broken version of the protected link controller for this issue. I have placed the pieces of code relevant to this issue/story (the database portion) into the old, working, unprotected LinkController from the master branch. The protected LinkController issue will have to be fixed separately (it already has its own issue/story and branch)